fixed the styling of two sets of messages; fixed the auto scroll to new messages when page loads and when new message comes
